+++ src/Enums/ToastAnimation.php
@@ -8,4 +8,8 @@ enum ToastAnimation: string
     case Slide = 'slide';
     case Scale = 'scale';
     case Spring = 'spring';
+    case Snap = 'snap';
+    case Pop = 'pop';
+    case Reveal = 'reveal';
+    case Bounce = 'bounce';
 }