+++ src/Exceptions/ToastKitException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Victorycodedev\ToastKit\Exceptions;
+
+use RuntimeException;
+
+class ToastKitException extends RuntimeException {}