+++ src/Facades/Toast.php
@@ -13,6 +13,8 @@ use Victorycodedev\ToastKit\PendingToastUpdate;
  * @method static PendingToast warning(string $message)
  * @method static PendingToast info(string $message)
  * @method static PendingToast neutral(string $message)
+ * @method static void definePreset(string $name, \Closure $preset)
+ * @method static PendingToast preset(string $name)
  * @method static PendingToastUpdate update(string $id)
  * @method static void dismiss(string $id)
  * @method static void dismissAll()